Rename package mustbe to assume
