+++ assume/assume.go
@@ -1,9 +1,9 @@
-// Package mustbe provides assertions that panic on violation.
+// Package assume provides assertions that panic on violation.
 //
 // This serves two purposes:
 //  1. Failing fast.
 //  2. Remove unnecessary branches and the temptation to test them.
-package mustbe
+package assume
 
 import "fmt"
 
@@ -18,8 +18,8 @@ func Equal[T comparable](x, y T) {
 	}
 }
 
-// NoError panics if the provided error is not nil.
-func NoError(err error) {
+// Success panics if the provided error is not nil.
+func Success(err error) {
 	if err != nil {
 		panic(fmt.Sprintf("unexpected error: %v", err))
 	}
